+++ Pi_Java_JCSP_MultiplexedChannel.java
@@ -2,7 +2,7 @@
  *  Calculation of Pi using quadrature realized with a fork/join approach with JCSP to partition the problem
  *  and hence harness all processors available to the JVM.
  *
- *  Copyright © 2009-10 Russel Winder
+ *  Copyright © 2009--2011 Russel Winder
  */
 
 import org.jcsp.lang.Channel ;
@@ -16,20 +16,20 @@ import java.util.List ;
 
 public class Pi_Java_JCSP_MultiplexedChannel {
   private static void execute ( final int numberOfTasks ) {
-    final long n = 1000000000l ;
+    final int n = 1000000000 ;
     final double delta = 1.0 / n ;
     final long startTimeNanos = System.nanoTime ( ) ;
-    final long sliceSize = n / numberOfTasks ;
+    final int sliceSize = n / numberOfTasks ;
     final Any2OneChannel<Double> channel = Channel.any2one ( ) ;
     final List<CSProcess> processes = new ArrayList<CSProcess> ( ) ;
     for ( int i = 0 ; i < numberOfTasks ; ++i ) {
       final int taskId = i ;
       processes.add ( new CSProcess ( ) {
           @Override public void run ( ) {
-            final long start = 1 + taskId * sliceSize ;
-            final long end = ( taskId + 1 ) * sliceSize ;
+            final int start = 1 + taskId * sliceSize ;
+            final int end = ( taskId + 1 ) * sliceSize ;
             double sum = 0.0 ;
-            for ( long i = start ; i <= end ; ++i ) {
+            for ( int i = start ; i <= end ; ++i ) {
               final double x = ( i - 0.5 ) * delta ;
               sum += 1.0 / ( 1.0 + x * x ) ;
             }
