+++ Pi_Java_ParallelArray.java
@@ -2,7 +2,7 @@
  *  Calculation of Pi using quadrature realized with an approached based on evolving values in an array with
  *  implicit parallelism.  This is material that probably won't actually be in Java 7.
  *
- *  Copyright © 2009-10 Russel Winder
+ *  Copyright © 2009--2011 Russel Winder
  */
 
 import extra166y.CommonOps ;
@@ -12,17 +12,17 @@ import extra166y.ParallelDoubleArrayWithDoubleMapping ;
 
 public class Pi_Java_ParallelArray {
   private static void execute ( final int numberOfTasks ) {
-    final long n = 1000000000l ;
+    final int n = 1000000000 ;
     final double delta = 1.0 / n ;
     final long startTimeNanos = System.nanoTime ( ) ;
-    final long sliceSize = n / numberOfTasks ;
+    final int sliceSize = n / numberOfTasks ;
     final ParallelDoubleArray initialArray = ParallelDoubleArray.create ( numberOfTasks , ParallelDoubleArray.defaultExecutor ( ) ) ;
     final ParallelDoubleArrayWithDoubleMapping finalArray = initialArray.withIndexedMapping ( new Ops.IntAndDoubleToDouble ( ) {
         @Override public double op ( final int taskId , final double initialValue ) {
-          final long start = 1 + taskId * sliceSize ;
-          final long end = ( taskId + 1 ) * sliceSize ;
-           double sum = 0.0 ;
-          for ( long i = start ; i <= end ; ++i ) {
+          final int start = 1 + taskId * sliceSize ;
+          final int end = ( taskId + 1 ) * sliceSize ;
+          double sum = 0.0 ;
+          for ( int i = start ; i <= end ; ++i ) {
             final double x = ( i - 0.5 ) * delta ;
             sum += 1.0 / ( 1.0 + x * x ) ;
           }
