+++ Pi_Java_Sequential.java
@@ -1,16 +1,16 @@
 /*
  *  Calculation of Pi using quadrature realized with a basic sequential algorithm.
  *
- *  Copyright © 2008-9 Russel Winder
+ *  Copyright © 2008--2011 Russel Winder
  */
 
 public class Pi_Java_Sequential {
   public static void main ( final String[] args ) {
-    final long n = 1000000000l ;
+    final int n = 1000000000 ;
     final double delta = 1.0 / n ;
     final long startTimeNanos = System.nanoTime ( ) ;
     double sum = 0.0 ;
-    for ( long i = 1 ; i <= n ; ++i ) {
+    for ( int i = 1 ; i <= n ; ++i ) {
       final double x = ( i - 0.5 ) * delta ;
       sum += 1.0 / ( 1.0 + x * x ) ;
     }
